Refactoring Unlogged and Logged pages; incrementing sign up form
