+++ web/src/pages/UnloggedArea/Login/Login.module.scss
@@ -1,4 +1,4 @@
-@import "../../variables.scss";
+@import "../../../variables.scss";
 
 // TODO: Responsivo
 .LoginFormContainer {
