+++ web/src/pages/UnloggedArea/UnloggedAreaLayout.module.scss
@@ -1,6 +1,6 @@
 @import '../../variables.scss';
 
-.UnloggedArea {
+.UnloggedAreaLayout {
   width: 100vw;
   height: 100vh;
   display: flex;
