+++ web/src/pages/UnloggedArea/index.tsx
@@ -1,23 +1,15 @@
 import React, { useEffect } from 'react';
 
-import UnloggedAreaRoutes from './routes';
+import classes from './UnloggedAreaLayout.module.scss';
 
-import classes from './UnloggedArea.module.scss';
-
-const UnloggedArea = (props: any) => {
-  useEffect(() => {
-    if (localStorage.getItem('token')) {
-      props.history.push('/');
-    }
-  }, []);
-
+const UnloggedAreaLayout = (props: any) => {
   return (
     <>
       {!localStorage.getItem('token') ? 
-        <div className={classes.UnloggedArea}>
+        <div className={classes.UnloggedAreaLayout}>
           <div className={classes.SidePanel}></div>
           <div className={classes.UnloggedAreaContent}>
-            <UnloggedAreaRoutes />
+            {props.children}
           </div>
         </div> : null
       }
@@ -25,4 +17,4 @@ const UnloggedArea = (props: any) => {
   )
 }
 
-export default UnloggedArea;
+export default UnloggedAreaLayout;
