+++ web/src/routes.tsx
@@ -4,14 +4,16 @@ import { BrowserRouter, Route } from 'react-router-dom';
 
 import AuthenticatedRoute from './authenticated-route';
 
-import LoggedArea from './pages/LoggedArea';
-import UnloggedArea from './pages/UnloggedArea';
+import LoggedAreaLayout from './pages/LoggedArea';
+import Login from './pages/UnloggedArea/Login';
+import SignUp from './pages/UnloggedArea/SignUp';
 
 const Routes = () => {
   return (
     <BrowserRouter>
-      <Route path='/' component={UnloggedArea} />
-      <AuthenticatedRoute path='/' component={LoggedArea} />
+      <Route path='/login' component={Login} />
+      <Route path='/cadastro' component={SignUp} />
+      <AuthenticatedRoute path='/' component={LoggedAreaLayout} />
     </BrowserRouter>
   )
 }
